Added error handling  + array rework
Realized that having a value in the global array and removing it each time was useless, no clue why I had that added in the first place anyways since I just append the global array and there's no need to reset it since the program exits after each function. Error handling was implemented due to demand and just general ease of use. It's always better to say "Bad Input" or "Not an option" instead of having Python just print the error line because you typed 7 instead of 6.